User signup errors fixed | Confirmations for delete and logout added | Buttons on comments show up only if logged in

+++ frontend/src/components/Comment/index.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react"
 import { updateComment, deleteComment } from "../../../utils/backend"
 
-export default function Comment({data, refreshComments}){
+export default function Comment({data, refreshComments, loggedIn}){
     const [editMode, setEditMode] = useState(false)
     const [editModeData, setEditModeData]= useState({
         name: data.name,
@@ -22,17 +22,23 @@ export default function Comment({data, refreshComments}){
     }
 
     function handleDelete() {
-        deleteComment(data._id)
-            .then(() => refreshComments())
-            .catch(err=> console.log(err))
+        if(confirm("Are you sure you want to delete this comment?")){
+            deleteComment(data._id)
+                .then(() => refreshComments())
+                .catch(err=> console.log(err))
+        }
     }
 
     // Edit mode false
     let commentDisplay =<div className="border-t-4 border-white p-2 pl-5">
             <h1 className="text-lg md:text-xl">{data.name}</h1>
             <p className="text-md md:text-lg">{data.content}</p>
-            <button onClick={()=>{setEditMode(true)}} className="bg-emerald-100 rounded-lg w-20 text-md md:text-lg mr-2 shadow-lg hover:shadow-violet-300 hover:bg-emerald-300">Edit</button>
-            <button onClick={handleDelete} className="bg-emerald-100 rounded-lg w-20 text-md md:text-lg shadow-lg hover:shadow-violet-300 hover:bg-emerald-300">Delete</button>
+            {loggedIn?<>
+                <button onClick={()=>{setEditMode(true)}} className="bg-emerald-100 rounded-lg w-20 text-md md:text-lg mr-2 shadow-lg hover:shadow-violet-300 hover:bg-emerald-300">Edit</button>
+                <button onClick={handleDelete} className="bg-emerald-100 rounded-lg w-20 text-md md:text-lg shadow-lg hover:shadow-violet-300 hover:bg-emerald-300">Delete</button>
+                </>
+            :<></>
+            }
         </div>
 
     // Edit mode true
